requirements.txt
Django==1.11
djangorestframework==3.8.2
django-appconf==1.0.2
django-compressor==2.2
dj-database-url==0.5.0
pymongo==3.7.1
dj-static==0.0.6
gunicorn==19.9.0
static3==0.7.0
wsgiref==0.1.2
